refactor: 상수와 함수 분리
