+++ src/apis/instance/https.ts
@@ -1,8 +1,6 @@
-import axios, { AxiosRequestConfig, AxiosResponse, AxiosError } from "axios";
-import { postReissue } from "@/apis/common/token";
-import { baseInstance } from "./baseInstance";
-
-const RETRY_HEADER_KEY = "retry";
+import axios, { AxiosResponse, AxiosError } from "axios";
+import { canRetryRequest, getIsRefreshing, isTokenError, processQueue, refreshAccessToken, setIsRefreshing } from "@/functions/axios";
+import { REQUEST_QUEUE, RETRY_HEADER_KEY } from "@/constants/axios";
 
 export const https = axios.create({
   baseURL: `${process.env.NEXT_PUBLIC_BASE_URL}v1/`,
@@ -19,56 +17,6 @@ https.interceptors.request.use((config) => {
   return config;
 });
 
-interface QueuedRequest {
-  resolve: (value: AxiosResponse) => void;
-  reject: (reason: Error) => void;
-  config: AxiosRequestConfig;
-}
-
-const requestQueue: QueuedRequest[] = [];
-
-let isRefreshing = false;
-
-const isTokenError = (error: AxiosError): boolean => {
-  return error.response?.status === 401;
-};
-
-const canRetryRequest = (config: AxiosRequestConfig | undefined): boolean => {
-  return config !== undefined && !config.headers?.[RETRY_HEADER_KEY];
-};
-
-const refreshAccessToken = async (): Promise<string> => {
-  const result = await postReissue();
-  if (!result) {
-    throw new Error("토큰 갱신에 실패했습니다.");
-  }
-  return result.accessToken;
-};
-
-const processQueue = (error: Error | null, token: string | null = null): void => {
-  requestQueue.forEach(({ resolve, reject, config }) => {
-    if (error) {
-      reject(error);
-      return;
-    }
-    if (token) {
-      const configCopy: AxiosRequestConfig = {
-        ...config,
-        headers: {
-          ...config.headers,
-          Authorization: `Bearer ${token}`,
-        },
-      };
-      if (configCopy.headers) {
-        delete configCopy.headers[RETRY_HEADER_KEY];
-      }
-      baseInstance(configCopy).then(resolve).catch(reject);
-    }
-  });
-  requestQueue.length = 0;
-  isRefreshing = false;
-};
-
 https.interceptors.response.use(
   (response: AxiosResponse) => response,
   async (error: AxiosError) => {
@@ -82,13 +30,13 @@ https.interceptors.response.use(
     }
 
     return new Promise<AxiosResponse>((resolve, reject) => {
-      requestQueue.push({ resolve, reject, config: originalRequest! });
+      REQUEST_QUEUE.push({ resolve, reject, config: originalRequest! });
 
-      if (isRefreshing) {
+      if (getIsRefreshing()) {
         return;
       }
 
-      isRefreshing = true;
+      setIsRefreshing(true);
 
       refreshAccessToken()
         .then((newAccessToken) => {
